+++ app/demo/flvfile2es/flvfile2es.go
@@ -31,6 +31,7 @@ func main() {
 		option.AssertBehavior = nazalog.AssertFatal
 	})
 	defer nazalog.Sync()
+	base.LogoutStartInfo()
 
 	var err error
 	flvFileName, aacFileName, avcFileName := parseFlag()