+++ app/demo/pullrtmp/pullrtmp.go
@@ -52,6 +52,7 @@ func main() {
 		option.AssertBehavior = nazalog.AssertFatal
 	})
 	defer nazalog.Sync()
+	base.LogoutStartInfo()
 
 	urlTmpl, fileNameTmpl, num := parseFlag()
 	nazalog.Infof("parse flag succ. urlTmpl=%s, fileNameTmpl=%s, num=%d", urlTmpl, fileNameTmpl, num)